Move fixtures to separate conftest.py file
The goal will be to reuse the fixture for other tests
