+++ test/test_map.py
@@ -1,33 +1,4 @@
-import pytest
-from qgis.utils import plugins
-from PyQt5.QtCore import QSettings, QCoreApplication
-
 from ee_plugin import Map
-from ee_plugin.ee_plugin import GoogleEarthEnginePlugin
-
-
-@pytest.fixture(scope="module", autouse=True)
-def setup_ee():
-    """Initialize the Earth Engine API."""
-    import ee
-
-    ee.Initialize()
-    ee.Authenticate(auth_mode="localhost", quiet=True)
-
-
-@pytest.fixture(scope="module", autouse=True)
-def load_ee_plugin(qgis_app):
-    """Load Earth Engine plugin and configure QSettings."""
-
-    # Set QSettings values required by the plugin
-    QCoreApplication.setOrganizationName("QGIS")
-    QCoreApplication.setApplicationName("QGIS Testing")
-    QSettings().setValue("locale/userLocale", "en")
-
-    # Initialize and register the plugin
-    plugin = GoogleEarthEnginePlugin(qgis_app)
-    plugins["ee_plugin"] = plugin
-    plugin.check_version()
 
 
 def test_add_layer():
